Add ComputerControllerTest along with tests

+++ src/main/kotlin/models/Computer.kt
@@ -1,8 +1,8 @@
 package models
 
-data class Computer (
+data class Computer(
     var computerId: Int,
     var computerName: String,
     var operatingSystem: String,
-    var storage: Float
+    var storage: Double
 )
